refactor: move editor into description-editor component
